requirements.txt
# RAG核心依赖
langchain
langchain-community
langchain-openai

# 文档处理
pandas
openpyxl
pypdf2
docx2txt

# 向量存储
faiss-cpu

# Web界面
streamlit

# FastAPI框架
fastapi
uvicorn[standard]
python-multipart
websockets

# 配置管理
python-dotenv

# API客户端
openai

# 数据验证
pydantic
pydantic-settings

# HTTP客户端
httpx
aiofiles

# 认证和安全
python-jose[cryptography]
passlib[bcrypt]

# 数据库（可选）
sqlalchemy
alembic

# 监控和日志
prometheus-client
structlog

# 测试
pytest
pytest-asyncio
httpx
